Create Exception Handling
The Datax shipping company wants to calculate the average cost of each item in the container. Get the container price and the number of items in the container as input. Write a program to get the average cost of each item. This program will throw an arithmetic exception when number of items in the container is zero(divide by zero). Use Exception handling mechanism to handle the exception. Use single try and catch block. In catch block print the exception thrown. 

Problem Specification :
The file name should be Main.java

Input format :
The first integer consists of an integer which corresponds to the number of items in the container. 
The second integer consists of an integer which corresponds to the total amount of items in the container. 

Output format :
The output consists of an Integer value which corresponds to the average cost of the item.
If the number of items in the container is 0 it will throw an exception, print the exception in the specified format.
Refer sample input and output for display specifications.

[All text in bold corresponds to input and rest corresponds to output]
Sample Input and Output 1: 
Enter the container price :
12000
Enter the number of items in the container :
 0
Exception : java.lang.ArithmeticException

Sample Input and Output 2: 
Enter the container price :
35000 
Enter the number of items in the container :
500
The average price of the item is Rs.70